requirements.txt
asgiref==3.4.0
Django==3.2.5
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.4